+++ scripts/global.js
@@ -2,7 +2,7 @@
     is it really a script if it's just holding a global variable? */
 
 const watermark = 'Kival Evan#5480';
-const version = 'v1.5.0';
+const version = 'v1.5.1';
 
 const flag = {
     noAudio: false,