+++ scripts/index.js
@@ -144,12 +144,12 @@ function printHTMLBold(arg1, arg2) {
     return `<b>${arg1}</b>: ${arg2.join(', ')}`;
 }
 
-function isAboveTH(t, th) {
-    return toRealTime(t) > th;
+function isAboveThres(t, rt) {
+    return toRealTime(t) > rt;
 }
 
-function isBelowTH(t, th) {
-    return toRealTime(t) < th;
+function isBelowThres(t, rt) {
+    return toRealTime(t) < rt;
 }
 
 function compToHex(c) {